Preserve old functionality
Previously, we were trying to maintain two sets of functionality
simultaniously. Since the old API was not tested, it's difficult to tell
if we have broken anything, and we have to add extra code so that we can
function correctly for both sets of functionality. This commit restores
the old functionality (mostly) verbatim. Since it's separated into a new
module, we can a) easily tell what has been ported, and b) don't have to
worry about breaking old code.