Update the users on a group properly
Previously, if you added a user to the group, it would only keep the users that
were on the group passed to the function. If you done this:

```
group = create_group
add_user(group, first_user)
add_user(group, second_user)
group = group_show
```

The final `group.users` would only include the `second_user` and not the first.

This resolves the issue by treating the `group.users` from the server as the
single source of truth.