Ensure that the Fake bundle performs as the HTTP
`HTTP.Bundle` runs `Bundle.encode_status`
which did not accept a string.
This meant that if you passed a string,
the fake would work fine,
but `HTTP.Bundle` would throw an exception.
This ensures that we use the same function
so the fake would have blown up as well.
It also coerces the string into a boolean.