Store associations as references
Storing as full objects will work fine about 90% of the time the problem comes in a scenario like this:

* Create a relay group
* Add a relay called "bob" to that relay group
* All is well and good!

Then:

* Change the name of that relay to be "bill" instead of "bob"

Now, if you go to "edit" the relay, it will be named properly but when you list all the relays for that relay group, there is actually a copy of the relay but it is still named "bill".

This changes to a pattern where we can store the associations as ids and then expand them when they're pulled out of the server. This is an "opt-in" system and I only opted in on the one place we need them for now.

I also changed to using a fake_server_information. I think this makes it a bit easier and you need to know less about the server when making calls and instead, the resource knows how to handle that. It also gives us a place to define those relationships that I'm certain we'll want to utilize more moving forward.